+++ source/feature-manager.tsx
@@ -27,6 +27,7 @@ import {
 } from './helpers/hotfix.js';
 import asyncForEach from './helpers/async-for-each.js';
 import {catchErrors, disableErrorLogging} from './helpers/errors.js';
+import {getFeatureID, log, shortcutMap} from './helpers/feature-helpers.js';
 
 type FeatureInitResult = void | false;
 type FeatureInit = (signal: AbortSignal) => Promisable<FeatureInitResult>;
@@ -53,21 +54,12 @@ type InternalRunConfig = RunConditions & {
 	shortcuts: Record<string, string>;
 };
 
-const shortcutMap = new Map<string, string>();
-const getFeatureID = (url: string): FeatureID => url.split('/').pop()!.split('.')[0] as FeatureID;
-
 const currentFeatureControllers = new ArrayMap<FeatureID, AbortController>();
-
-const log = {
-	info: console.log,
-	http: console.log.bind(console, '🌏'),
-};
 
 // eslint-disable-next-line no-async-promise-executor -- Rule assumes we don't want to leave it pending
 const globalReady = new Promise<RGHOptions>(async resolve => {
-	// This file may be imported in the options
 	if (!isWebPage()) {
-		return;
+		throw new Error('This script should only be run on web pages');
 	}
 
 	const [options, localHotfixes, bisectedFeatures] = await Promise.all([
@@ -76,6 +68,8 @@ const globalReady = new Promise<RGHOptions>(async resolve => {
 		bisectFeatures(),
 		preloadSyncLocalStrings(),
 	]);
+
+	log.setup(options);
 
 	await waitFor(() => document.body);
 
@@ -112,15 +106,6 @@ const globalReady = new Promise<RGHOptions>(async resolve => {
 		// If features are remotely marked as "seriously breaking" by the maintainers, disable them without having to wait for proper updates to propagate #3529
 		void brokenFeatures.get();
 		Object.assign(options, localHotfixes);
-	}
-
-	// Create logging function
-	if (!options.logging) {
-		log.info = () => {/* No logging */};
-	}
-
-	if (!options.logHTTP) {
-		log.http = () => {/* No logging */};
 	}
 
 	if (elementExists('body.logged-out')) {
@@ -161,26 +146,6 @@ async function maybeRun(id: FeatureID, {asLongAs, include, exclude, init, shortc
 			}
 		}
 	});
-}
-
-type FeatureHelper = {
-	/** If `import.meta.url` is passed as URL, this will be the feature ID */
-	id: string;
-
-	/** A class name that can be added as attribute */
-	class: string;
-
-	/** A class selector that can be used with querySelector */
-	selector: string;
-};
-
-function getIdentifiers(url: string): FeatureHelper {
-	const id = getFeatureID(url);
-	return {
-		id,
-		class: 'rgh-' + id,
-		selector: '.rgh-' + id,
-	};
 }
 
 /** Register a new feature */
@@ -287,10 +252,6 @@ const features = {
 	add,
 	unload,
 	addCssFeature,
-	log,
-	shortcutMap,
-	getFeatureID,
-	getIdentifiers,
 };
 
 export default features;
