+++ source/helpers/fetch-dom.ts
@@ -2,12 +2,12 @@ import mem from 'memoize';
 import domify from 'doma';
 import type {ParseSelector} from 'typed-query-selector/parser.js';
 
-import features from '../feature-manager.js';
+import {log} from './feature-helpers.js';
 
 async function fetchDom(url: string): Promise<DocumentFragment>;
 async function fetchDom<Selector extends string, TElement extends HTMLElement = ParseSelector<Selector, HTMLElement>>(url: string, selector: Selector): Promise<TElement | undefined>;
 async function fetchDom(url: string, selector?: string): Promise<Node | undefined> {
-	features.log.http(url);
+	log.http(url);
 	const absoluteURL = new URL(url, location.origin).href; // Firefox `fetch`es from the content script, so relative URLs fail
 	const response = await fetch(absoluteURL);
 	const dom = domify(await response.text());
